requirements.txt
tabula-py==2.9.3
pandas==2.2.2
JPype1==1.5.0
Django==5.1.1
django-tailwind==3.8.0
django-browser-reload==1.15.0
langchain_core==0.2.38
langchain-google-genai==1.0.10
python-dotenv==1.0.1
langchain==0.2.16
langchain-community==0.2.16
langchain-chroma==0.1.3
jq==1.8.0